added float nums

+++ app.js
@@ -8,8 +8,8 @@ app.listen(8000, function() {
 // WHEN a user goes to http://localhost:3000/add/num/num
 app.get("/add/:numOne/:numTwo", function (req, res) {
   // setting each url property to a integer and assigned to var
-  var numOne = parseInt(req.params.numOne); 
-  var numTwo = parseInt(req.params.numTwo);
+  var numOne = parseFloat(req.params.numOne); 
+  var numTwo = parseFloat(req.params.numTwo);
   //add vars
   var total = numOne + numTwo;
   //send the total back
@@ -19,8 +19,8 @@ app.get("/add/:numOne/:numTwo", function (req, res) {
 // WHEN a user goes to http://localhost:3000/sub/num/num
 app.get("/sub/:numOne/:numTwo", function (req, res) {
   // setting each url property to a integer and assigned to var
-  var numOne = parseInt(req.params.numOne); 
-  var numTwo = parseInt(req.params.numTwo);
+  var numOne = parseFloat(req.params.numOne); 
+  var numTwo = parseFloat(req.params.numTwo);
   //add vars
   var total = numOne - numTwo;
   //send the total back
@@ -30,8 +30,8 @@ app.get("/sub/:numOne/:numTwo", function (req, res) {
 // WHEN a user goes to http://localhost:3000/multi/num/num
 app.get("/multi/:numOne/:numTwo", function (req, res) {
   // setting each url property to a integer and assigned to var
-  var numOne = parseInt(req.params.numOne); 
-  var numTwo = parseInt(req.params.numTwo);
+  var numOne = parseFloat(req.params.numOne); 
+  var numTwo = parseFloat(req.params.numTwo);
   //add vars
   var total = numOne * numTwo;
   //send the total back
@@ -41,8 +41,8 @@ app.get("/multi/:numOne/:numTwo", function (req, res) {
 // WHEN a user goes to http://localhost:3000/div/num/num
 app.get("/div/:numOne/:numTwo", function (req, res) {
   // setting each url property to a integer and assigned to var
-  var numOne = parseInt(req.params.numOne); 
-  var numTwo = parseInt(req.params.numTwo);
+  var numOne = parseFloat(req.params.numOne); 
+  var numTwo = parseFloat(req.params.numTwo);
   //add vars
   var total = numOne / numTwo;
   //send the total back
